refactor(test): split two files back under the 300-line bloat gate
Pure extraction, no behavior change:
- _test_gate_fidelity.py (306 lines) -> moved _categorize_fidelity_screen
  and its two triage-key constants into a new _test_gate_fidelity_categorize.py,
  re-exported under their original names so no call site or test import
  changes. 270 lines.
- test_test_gate_e2e_reconciliation.py (303 lines) -> moved the three
  symlink-escape regression tests added in the prior commit into a new
  test_test_gate_e2e_reconciliation_hardening.py, sharing the existing
  fixture helpers via relative import (same split pattern already used
  for test_test_gate_fidelity_triage_hardening.py). 238 lines.

Verified: 258/258 in the gate/ledger test filter, ruff clean.
